+++ src/Stack/Types/Docker.hs
@@ -5,6 +5,7 @@
 module Stack.Types.Docker where
 
 import Control.Applicative
+import Control.Monad.Catch (MonadThrow)
 import Data.Aeson.Extended
 import Data.Monoid
 import Data.Text (Text)
@@ -40,6 +41,8 @@ data DockerOpts = DockerOpts
     -- ^ Environment variables to set in the container.
   ,dockerDatabasePath :: !(Path Abs File)
     -- ^ Location of image usage database.
+  ,dockerStackExe :: !DockerStackExe
+    -- ^ Location of container-compatible stack executable
   }
   deriving (Show)
 
@@ -76,6 +79,8 @@ data DockerOptsMonoid = DockerOptsMonoid
     -- ^ Environment variables to set in the container
   ,dockerMonoidDatabasePath :: !(Maybe String)
     -- ^ Location of image usage database.
+  ,dockerMonoidStackExe :: !(Maybe String)
+    -- ^ Location of container-compatible stack executable
   }
   deriving (Show)
 
@@ -98,6 +103,7 @@ instance FromJSON (DockerOptsMonoid, [JSONWarning]) where
               dockerMonoidMount            <- o ..:? dockerMountArgName ..!= []
               dockerMonoidEnv              <- o ..:? dockerEnvArgName ..!= []
               dockerMonoidDatabasePath     <- o ..:? dockerDatabasePathArgName
+              dockerMonoidStackExe         <- o ..:? dockerStackExeArgName
               return DockerOptsMonoid{..})
 
 -- | Left-biased combine Docker options
@@ -117,6 +123,7 @@ instance Monoid DockerOptsMonoid where
     ,dockerMonoidMount            = []
     ,dockerMonoidEnv              = []
     ,dockerMonoidDatabasePath     = Nothing
+    ,dockerMonoidStackExe         = Nothing
     }
   mappend l r = DockerOptsMonoid
     {dockerMonoidExists           = dockerMonoidExists l <|> dockerMonoidExists r
@@ -133,7 +140,24 @@ instance Monoid DockerOptsMonoid where
     ,dockerMonoidMount            = dockerMonoidMount r <> dockerMonoidMount l
     ,dockerMonoidEnv              = dockerMonoidEnv r <> dockerMonoidEnv l
     ,dockerMonoidDatabasePath     = dockerMonoidDatabasePath l <|> dockerMonoidDatabasePath r
+    ,dockerMonoidStackExe         = dockerMonoidStackExe l <|> dockerMonoidStackExe r
     }
+
+-- | Where to get the `stack` executable to run in Docker containers
+data DockerStackExe
+    = DockerStackExeDownload  -- ^ Download from official bindist
+    | DockerStackExeHost  -- ^ Host's `stack` (linux-x86_64 only)
+    | DockerStackExeImage  -- ^ Docker image's `stack` (versions must match)
+    | DockerStackExePath (Path Abs File) -- ^ Executable at given path
+    deriving (Show)
+
+-- | Parse 'DockerStackExe'.
+parseDockerStackExe :: MonadThrow m => String -> m DockerStackExe
+parseDockerStackExe t
+    | t == dockerStackExeDownloadVal = return DockerStackExeDownload
+    | t == dockerStackExeHostVal = return DockerStackExeHost
+    | t == dockerStackExeImageVal = return DockerStackExeImage
+    | otherwise = fmap DockerStackExePath (parseAbsFile t)
 
 -- | Docker volume mount.
 data Mount = Mount String String
@@ -217,3 +241,19 @@ dockerPersistArgName = "persist"
 -- | Docker database path argument name.
 dockerDatabasePathArgName :: Text
 dockerDatabasePathArgName = "database-path"
+
+-- | Docker database path argument name.
+dockerStackExeArgName :: Text
+dockerStackExeArgName = "stack-exe"
+
+-- | Value for @--docker-stack-exe=download@
+dockerStackExeDownloadVal :: String
+dockerStackExeDownloadVal = "download"
+
+-- | Value for @--docker-stack-exe=host@
+dockerStackExeHostVal :: String
+dockerStackExeHostVal = "host"
+
+-- | Value for @--docker-stack-exe=image@
+dockerStackExeImageVal :: String
+dockerStackExeImageVal = "image"